exam with other changes

+++ server/models/user.js
@@ -12,6 +12,9 @@ module.exports= {
   getUserInfo(id) {
     return conn.query('select id, first_name, last_name, nickname, birthday, user_icon from users where id=?', id)
   },
+  getUserInfo2(id) {
+    return conn.query(`select id, first_name, last_name, nickname, birthday, user_icon from users where id REGEXP '^${id}'`)
+  },
   //get friend posts
   getUserRecords(id) {
     return conn.query('select records.id, records.created_at, type, amount from records join exercise on records.exer_id = exercise.id where user_id = ?;',
@@ -19,16 +22,23 @@ module.exports= {
   },
   //get thumb-up and comments related to post
   getUserRecordsCT(id) {
-    return conn.query(`select thumbup_user_id,records_id from thumbup_to_records where records_id = any(select records.id from records where user_id = ?);
-      select comments_to_records.id comment_user_id, comment_text, records_id from comments_to_records where records_id = any(select records.id from records where user_id = ?);`,
+    return conn.query(`select thumbup_user_id, records_id from thumbup_to_records where records_id = any(select records.id from records where user_id = ?);
+      select comment_user_id, comment_user_firstname, comment_text, records_id from comments_to_records where records_id = any(select records.id from records where user_id = ?);`,
       [id, id])
   },
   addUserRecordsCT(input) {
-    return conn.query('insert into thumbup_to_records (created_at, records_id, thumbup_user_id) value(?)',
-    [[new Date(), input.records_id, input.user_id]])
+    if (input.type === 'thumbUp')
+      return conn.query('insert into thumbup_to_records (created_at, records_id, thumbup_user_id) value(?)',
+      [[new Date(), input.records_id, input.user_id]])
+    if (input.type === 'comment')
+      return conn.query('insert into comments_to_records (created_at, records_id, comment_user_id, comment_text, comment_user_firstname) value(?)',
+      [[new Date(), input.records_id, input.comment_user_id, input.comment_text, input.comment_user_firstname]])
   },
   deleteUserRecordsCT(input) {
-    return conn.query(`delete from thumbup_to_records where records_id = ? and thumbup_user_id = ?;`, [input.targetRecord, input.targetUser])
+    if (input.type === 'thumbUp')
+      return conn.query(`delete from thumbup_to_records where records_id = ? and thumbup_user_id = ?;`, [input.targetRecord, input.targetUser])
+    if (input.type === 'comment') 
+      return conn.query(`delete from comments_to_records where records_id = ? and comment_user_id = ?;`, [input.targetRecord, input.targetUser])
   },
   //after login in, grab information about the user
   getInfo(id) {
